Increase no. of jobs scrapped.
