--- output/jobhai_driver_jobs.csv
@@ -1,29 +1,0 @@
-Job Title,Company Name,Experience,Salary,Location,URL
-Cab Driver,"Everest FleetSingasandra, Bangalore (Field job)Skills:Automatic Car Driving, Cab Driving",3 years,₹45000,Bangalore,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-singasandra-bangalore-0-to-3-years-1760780990-6712425-jid
-Cab Driver,N/A,N/A,"Cab Driver ₹ 30,000 - 45,000 per month *",N/A,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-singasandra-bangalore-0-to-3-years-1760780990-6712425-jid
-N/A,N/A,3 years,₹45000,Bangalore,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-singasandra-bangalore-0-to-3-years-1760780990-6712425-jid
-Cab Driver,"Everest FleetTaratala, Kolkata (Field job)Skills:Cab Driving",N/A,"Cab Driver ₹ 30,000 - 45,000 per month *",Kolkata,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-taratala-kolkata-0-to-3-years-1760780972-6712243-jid
-Ola / Uber Driver,"Everest FleetThane West, Thane (Field job)Skills:Cab Driving",3 years,₹45000,Mumbai,https://www.jobhai.com/driver-ola-uber-driver-job-in-everest-fleet-private-limited-thane-west-mumbai-0-to-3-years-1760780957-6712165-jid
-Ola / Uber Driver,N/A,N/A,"Ola / Uber Driver ₹ 30,000 - 45,000 per month *",N/A,https://www.jobhai.com/driver-ola-uber-driver-job-in-everest-fleet-private-limited-thane-west-mumbai-0-to-3-years-1760780957-6712165-jid
-N/A,N/A,3 years,₹45000,Mumbai,https://www.jobhai.com/driver-ola-uber-driver-job-in-everest-fleet-private-limited-thane-west-mumbai-0-to-3-years-1760780957-6712165-jid
-Cab Driver,"Everest FleetSector 10 Dwarka, Delhi (Field job)Skills:Cab Driving",3 years,₹45000,Delhi,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-sector-10-dwarka-delhi-0-to-3-years-1760775242-6711889-jid
-N/A,N/A,3 years,₹45000,Delhi,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-sector-10-dwarka-delhi-0-to-3-years-1760775242-6711889-jid
-Cab Driver,"Everest FleetBaner, Pune (Field job)Skills:Cab Driving, 4-Wheeler Driving Licence, Aadhar Card, Automatic Car Driving, PAN Card",3 years,₹45000,Pune,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-baner-pune-0-to-3-years-1760769332-6711615-jid
-N/A,N/A,3 years,₹45000,Pune,https://www.jobhai.com/driver-cab-driver-job-in-everest-fleet-private-limited-baner-pune-0-to-3-years-1760769332-6711615-jid
-Driver,"Everest FleetSector 9 Rohini, Delhi (Field job)Skills:Aadhar Card, PAN Card, Private Car Driving, Cab Driving, Bank Account, Smartphone, Auto/Tempo Driving, 4-Wheeler Driving Licence, Luxury Car Driving",N/A,"Driver ₹ 35,000 - 45,000 per month",Delhi,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-sector-9-rohini-delhi-0-to-6-plus-years-1760764950-6711252-jid
-Driver,N/A,N/A,"Driver ₹ 35,000 - 45,000 per month",N/A,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-sector-9-rohini-delhi-0-to-6-plus-years-1760764950-6711252-jid
-Driver,"Everest FleetSector 10 Rohini, Delhi (Field job)Skills:PAN Card, Private Car Driving, Smartphone, RC, Auto/Tempo Driving, Aadhar Card, Bank Account, Luxury Car Driving, 4-Wheeler Driving Licence",6+ months,₹45000 per month,Delhi,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-sector-10-rohini-delhi-0-to-6-plus-years-1760763959-6711128-jid
-Driver,"Everest FleetSector 6 Dwarka, Delhi (Field job)Skills:Auto/Tempo Driving, Private Car Driving, Luxury Car Driving, Smartphone, PAN Card, Bank Account, 4-Wheeler Driving Licence, Aadhar Card",N/A,"Driver ₹ 35,000 - 40,000 per month",Delhi,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-sector-6-dwarka-delhi-0-to-6-plus-years-1760763201-6710944-jid
-Driver,"Everest FleetB1 Block Sector 57 Gurgaon, Gurgaon (Field job)Skills:Automatic Car Driving, Aadhar Card, Bank Account, Smartphone, Private Car Driving, 4-Wheeler Driving Licence, Cab Driving, Luxury Car Driving, Auto/Tempo Driving, PAN Card",N/A,"Driver ₹ 30,000 - 45,000 per month",Gurgaon,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-b1-block-sector-57-gurgaon-gurgaon-0-to-6-plus-years-1760763196-6710925-jid
-Driver,"Everest FleetNajafgarh, Delhi (Field job)Skills:4-Wheeler Driving Licence, Cab Driving, Private Car Driving, Bank Account, Auto/Tempo Driving, PAN Card, Automatic Car Driving, Luxury Car Driving, Aadhar Card, Smartphone",72 months,₹55000,Delhi,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-najafgarh-delhi-0-to-6-years-1760677828-6706981-jid
-Ola / Uber Driver,"Shree Krishna Tours And TravelsVasai, Mumbai (Field job)0 - 6+ years Experience in Driver",6+ years,₹40000,Mumbai,https://www.jobhai.com/driver-ola-uber-driver-job-in-shree-krishna-tours-and-travels-vasai-mumbai-0-to-6-plus-years-1760890462-6713711-jid
-Ola / Uber Driver,"Shree Krishna Tours And TravelsVirar, Mumbai (Field job)0 - 6+ years Experience in Driver",6+ years,₹40000,Mumbai,https://www.jobhai.com/driver-ola-uber-driver-job-in-shree-krishna-tours-and-travels-virar-mumbai-0-to-6-plus-years-1760886642-6713690-jid
-Ola / Uber Driver,"Shree Krishna Tours And TravelsNallasopara, Mumbai (Field job)0 - 6+ years Experience in Driver",6+ years,"Ola / Uber Driver ₹ 30,000 - 40,000 per month *",Mumbai,https://www.jobhai.com/driver-ola-uber-driver-job-in-shree-krishna-tours-and-travels-nallasopara-mumbai-0-to-6-plus-years-1760886522-6713689-jid
-Driver,"Everest FleetBlock C1 sector 3 Gurgaon, Gurgaon (Field job)Skills:Auto/Tempo Driving, Cab Driving, PAN Card, Private Car Driving, Smartphone, Aadhar Card, Bank Account, Luxury Car Driving, 4-Wheeler Driving Licence",N/A,"Driver ₹ 30,000 - 40,000 per month",Gurgaon,https://www.jobhai.com/driver-driver-job-in-everest-fleet-private-limited-block-c1-sector-3-gurgaon-gurgaon-0-to-6-plus-years-1760775436-6711585-jid
-Driver,"Everest Fleetkhairatabad, Hyderabad0 - 6 years Experience in Driver",6 years,₹80000,Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-khairatabad-hyderabad-0-to-6-years-1760528403-6702207-jid
-N/A,N/A,6 years,₹80000,Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-khairatabad-hyderabad-0-to-6-years-1760528403-6702207-jid
-Driver,"Everest FleetSuraram, Hyderabad0 - 6 years Experience in Driver",6 years,"Driver ₹ 70,000 - 80,000 per month",Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-suraram-hyderabad-0-to-6-years-1760528403-6702206-jid
-Driver,"Everest FleetKothapet, Hyderabad0 - 6 years Experience in Driver",6 years,₹80000,Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-kothapet-hyderabad-0-to-6-years-1760528403-6702205-jid
-Driver,"Everest FleetPuranapool, Hyderabad0 - 6 years Experience in Driver",6 years,₹80000 per month,Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-puranapool-hyderabad-0-to-6-years-1760528403-6702204-jid
-Driver,"Everest FleetMailardevpally, Hyderabad0 - 6 years Experience in Driver",6 years,"Driver ₹ 70,000 - 80,000 per month",Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-mailardevpally-hyderabad-0-to-6-years-1760528403-6702203-jid
-Driver,"Everest FleetEdi Bazaar, Hyderabad0 - 6 years Experience in Driver",6 years,₹80000 per month,Hyderabad,https://www.jobhai.com/driver-driver-job-in-everest-fleet-edi-bazaar-hyderabad-0-to-6-years-1760528403-6702202-jid
